+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.37
 
+- Add Simulate App For Simulate potrait and landscape at same time
 - update package